+++ ClassLibrary/Sdp/SdpAnswerSettings.cs
@@ -63,8 +63,13 @@ public class SdpAnswerSettings
     public string UserName { get; set; }
 
     /// <summary>
-    /// Fingerprint of the self-signed X.509 certificate that that will be used for DTLS-SDES keying
-    /// material negotiation
+    /// Fingerprint of the self-signed X.509 certificate that that will be used for DTLS-SRTP keying
+    /// material negotiation.
+    /// <para>
+    /// The RtpChannel class automatically creates a static X.509 certificate object that it will use for the
+    /// DTLS handshake if a media channel uses DTLS-SRTP encryption. It is best to pass the static RtpChannel.CertificateFingerprint property
+    /// for this property.
+    /// </para>
     /// </summary>
     /// <value></value>
     public string Fingerprint { get; set; }
@@ -82,7 +87,10 @@ public class SdpAnswerSettings
     /// <param name="VideoCodecs">List of supported video codecs</param>
     /// <param name="userName">User name to use for the session owner and MSRP URI in the media descriptions</param>
     /// <param name="fingerprint">Fingerprint of the self-signed X.509 certificate that will be used for DTLS-SDES keying
-    /// material negotiation</param>
+    /// material negotiation. The RtpChannel class automatically creates a static X.509 certificate object that it will use for the
+    /// DTLS handshake if a media channel uses DTLS-SRTP encryption. It is best to pass the static RtpChannel.CertificateFingerprint property
+    /// for this parameter.
+    /// </param>
     /// <param name="portManager">MediaPortManager to use use for allocation of media ports.</param>
     public SdpAnswerSettings(List<string> AudioCodecs, List<string> VideoCodecs, string userName, string fingerprint, 
         IMediaPortManager portManager)